fix(frontend): add `rel="external"` to external links
This patch resolves the CSRF issue with some apps (most notably, QBitTorrent).
TL;DR: some apps enforce that the `Referer` and `Target` headers must match,
which isn't the case by default with `a` tags, `Referer` is the URL you were
redirected from instead.

This is generally fixed with `rel="noopener noreferrer"` (as implemented by
@toboshii) but SvelteKit's router renders those options useless unless
`rel="external"` is set as well.